+++ api/routes/auth.js
@@ -1,6 +1,7 @@
 import express from 'express'
-import { register, login, logout } from '../controllers/auth.js'
+import { register, login, logout, search } from '../controllers/auth.js'
 import { csrfToken } from '../controllers/form.js'
+import { verifyAdmin } from '../utils/verifyToken.js'
 
 const router = express.Router()
 
@@ -9,5 +10,6 @@ router.post('/register', register)
 router.get('/login', csrfToken)
 router.post('/login', login)
 router.get('/logout', logout)
+router.get('/search', verifyAdmin, search)
 
 export default router