+++ client/src/components/Sidebar.jsx
@@ -7,6 +7,7 @@ import CalendarMonthIcon from '@mui/icons-material/CalendarMonth'
 import PlaceIcon from '@mui/icons-material/Place'
 import ComputerIcon from '@mui/icons-material/Computer'
 import PeopleIcon from '@mui/icons-material/People'
+import BookIcon from '@mui/icons-material/Book'
 import NoteAltIcon from '@mui/icons-material/NoteAlt'
 import FolderIcon from '@mui/icons-material/Folder'
 import { Link } from 'react-router-dom'
@@ -172,6 +173,12 @@ const itemDict =
             icon: <PeopleIcon/>,
             title: 'Employee',
             auth: true
+        },
+        {
+            to: '/loginhistory',
+            icon: <BookIcon/>,
+            title: 'LoginHistory',
+            auth: false
         },
         {
             to: '/board',