+++ client/src/configs/loginhistory.js
@@ -1,0 +1,69 @@
+export const columnHeaders = [
+    {
+        accessor: 'employeeId',
+        Header: 'ID',
+    },
+    {
+        accessor: 'name',
+        Header: 'Name',
+    },
+    {
+        accessor: 'date',
+        Header: 'Date',
+    },
+    {
+        accessor: 'time',
+        Header: 'Time',
+    },
+    {
+        accessor: 'ip',
+        Header: 'IP',
+    },
+    {
+        accessor: 'user_agent',
+        Header: 'USER_AGENT',
+    },
+    {
+        accessor: 'latitude',
+        Header: 'Latitude',
+    },
+    {
+        accessor: 'longitude',
+        Header: 'Longitude',
+    },
+]
+
+export const csvHeaders = [
+    {
+        key: 'name',
+        label: 'Name',
+    },
+    {
+        key: 'date',
+        label: 'Date',
+    },
+    {
+        key: 'time',
+        label: 'Time',
+    },
+    {
+        key: 'ip',
+        label: 'IP',
+    },
+    {
+        key: 'user_agent',
+        label: 'USER_AGENT',
+    },
+    {
+        key: 'location',
+        label: 'Location',
+    },
+    {
+        key: 'latitude',
+        label: 'Latitude',
+    },
+    {
+        key: 'longitude',
+        label: 'Longitude',
+    },
+]