+++ client/src/pages/TooManyRequests.jsx
@@ -1,0 +1,7 @@
+const TooManyRequests = () => {
+    return (
+        <h1>429 TooManyRequests</h1>
+    )
+}
+
+export default TooManyRequests